requirements.txt
alembic==1.4.2
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
crypto==1.4.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
flask-peewee==3.0.3
Flask-PW==1.1.3
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mock==4.0.2
Naked==0.1.31
peewee==3.13.3
peewee-migrate==1.1.6
pyactiveresource==2.2.1
pycrypto==2.6.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
ShopifyAPI==8.0.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
wtf-peewee==3.0.0
WTForms==2.3.1
